requirements.txt
beautifulsoup4
requests
selenium
clint